config.c: add configuration parameters for logging via mqtt.
Conceptually the logger will run in a low priority thread and communicate
every LogViaMQTTPeriodNS the values accumulated in a ring buffer via MQTT.
Here we also see the parameters for the MQTT connection.
MeasurementName will be used in a second stage in the influx as
Measurement Name, allowing several Linux Tsn Testbench running in
parallel.

Signed-off-by: Walfred Tedeschi <walfred.tedeschi@intel.com>