scripts: ehl: Adjust base time
Intel Elkhart Lake uses the stmmac which does support Qbv. The base time is
calculated to be 30s in the future. However, the PTP is based on
TAI. `date` is based on UTC. Meaning the difference is 37s. Therefore, use
date + 60s so that the base time is actually in the future.

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>