l2: Reduce number of system calls
Switch to recvmmsg() and sendmmsg() for AF_PACKET code to reduce the number
of required system calls. Especially when sending or receiving a lot of
frames the reduction of the system calls has a impact.

Example for 32 x 128 byte frames on Intel 10th generation CPU:

 - sendmsg()  - 88us
 - recvmsg()  - 41us
 - sendmmsg() - 69us
 - recvmmsg() - 31us

Tested on Linux v6.1-RT based on standard Debian configuration.

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>