tsn/rtc/rta: Introduce dedicated payload pattern for AE
The PROFINET encryption requires the plaintext of the frames
payload. Currently TxFrameData is abused for that. However, TxFrameData is
for all frames which are being sent and has basically nothing to do with
encryption.

The current code has several disadvantages:

 - TxFrameData has to be allocated even for XDP
 - Hinders to switch to sendmmsg()
 - Not obvious why allocate 2 instead of 1 frames

Therefore, decouple security and TxFrameData which are two different
things! And align g2 with tsn, rtc and rta.

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>