all: Implement clean shutdown
Currently the applications do not cleanup after finished. For instance, the
XDP program is left around. Therefore, implement a signal handler for
SIGTERM and SIGINT. This signal handler stops all the threads. Afterwards
the applications can cleanup.

Some threads use cond_wait(). Add a timeout so that the threads can check
the !stop condition. And last check the pids before calling
pthread_join(). In case a traffic class is not enabled all the thread ids
are zero.

Signed-off-by: Kurt Kanzenbach <kurt@linutronix.de>
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>